requirements.txt
requests
flask>=1.1
flask-login>=0.5
flask-wtf>=0.14
sqlalchemy>=1.3
sqlalchemy-serializer
flask-restful
bs4
python-telegram-bot
python-dotenv